+++ .env
@@ -1,0 +1,11 @@
+#베포 BASE 경로
+NEXT_PUBLIC_BASE_URL = "https://sp-globalnomad-api.vercel.app/6-11/"
+
+#구글 APP KEY 로그인
+NEXT_PUBLIC_GOOGLE_APP_KEY = "777040995318-03l3ia47hjduag0idnialt63vgv98fmq.apps.googleusercontent.com" 
+
+#카카오 APP KEY 로그인
+NEXT_PUBLIC_GOOGLE_APP_KEY = "9c116155dd69cb225ae77cf91935bcd5"
+
+#카카오 MAP KEY
+NEXT_PUBLIC_KAKAO_MAP_APP_KEY="1b091081a87df9961161821463277214"
