+++ app/components/MainPage/AllActivities/Activity.tsx
